nivel4/nivel4_app/requirements.txt
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
mysql-connector
redis==3.2.1
requests
